Moved itk geo building to module
